Fixed some bugs which made this simulation not runnable

+++ src/main/java/simulation/NetworkTrafficGenerator.java
@@ -4,5 +4,5 @@ package simulation;
  * Interface for generators of network traffic of VMs.
  */
 public interface NetworkTrafficGenerator {
-    public int generate(int previous, double max);
+    public int generate(int previous, int max);
 }
